Hip freshmake merge

+++ src/caffe/parallel.cpp
@@ -119,7 +119,7 @@ void DevicePair::compute(const vector<int> devices, vector<DevicePair>* pairs) {
   vector<int> remaining(devices);
 
   // Depth for reduction tree
-  int remaining_depth = static_cast<int>(ceil(log2(remaining.size())));
+  int remaining_depth = static_cast<int>(ceil(log2((float)remaining.size())));
 
   // Group GPUs by board
   for (int d = 0; d < remaining_depth; ++d) {
@@ -147,7 +147,7 @@ void DevicePair::compute(const vector<int> devices, vector<DevicePair>* pairs) {
   DLOG(INFO) << "GPUs paired by boards, remaining: " << s.str();
 
   // Group by P2P accessibility
-  remaining_depth = ceil(log2(remaining.size()));
+  remaining_depth = ceil(log2((float)remaining.size()));
   for (int d = 0; d < remaining_depth; ++d) {
     for (int i = 0; i < remaining.size(); ++i) {
       for (int j = i + 1; j < remaining.size(); ++j) {
@@ -170,7 +170,7 @@ void DevicePair::compute(const vector<int> devices, vector<DevicePair>* pairs) {
   DLOG(INFO) << "GPUs paired by P2P access, remaining: " << s.str();
 
   // Group remaining
-  remaining_depth = ceil(log2(remaining.size()));
+  remaining_depth = ceil(log2((float)remaining.size()));
   for (int d = 0; d < remaining_depth; ++d) {
     for (int i = 0; i < remaining.size(); ++i) {
       pairs->push_back(DevicePair(remaining[i], remaining[i + 1]));
